Add mover classes
