+++ Assets/Scripts/Actor/Movements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 249ec541b68f6974587d97ade55d46f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
