+++ Assets/Scripts/Movement/MoveSets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9adbe4fb7de24a6438ba60ce9a6f3031
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
