+++ Assets/Scripts/Movement/MoveSets/BasicMoveSet.cs
@@ -1,9 +1,8 @@
 ﻿using UnityEngine;
 
-// Moves object in a single direction. Great for testing movement.
 namespace Movement
 {
-    public class BasicMove : MonoBehaviour
+    public class BasicMoveSet : SlipTimeMover
     {
         public Vector3 moveDirection = new Vector3(0, 1, 0);
         public Vector3 velocity = new Vector3(1, 1, 1);
@@ -13,7 +12,7 @@ namespace Movement
         {
             Vector3 translateVect = moveDirection.normalized * Time.deltaTime;
             translateVect.Scale(velocity);
-            transform.position += translateVect;
+            this.SetMovement(translateVect);
         }
     }
 }
