+++ Assets/Scripts/Movement/MoveSets/README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a6be691fad136454aa1fa4363e3ff897
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
