+++ Assets/Scripts/Movement/SlipTimeMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3db522881b158824ea6b2c31ea118875
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
